Normalize matrix transforms across methods
Methods now match w3c expected results.
