+++ src/ImageSharp/Processing/Processors/Transforms/RotateProcessor.cs
@@ -47,7 +47,9 @@ namespace SixLabors.ImageSharp.Processing.Processors
         /// <inheritdoc/>
         protected override Matrix3x2 GetTransformMatrix()
         {
-            return Matrix3x2Extensions.CreateRotationDegrees(-this.Degrees, PointF.Empty);
+            Matrix3x2 matrix = Matrix3x2Extensions.CreateRotationDegrees(this.Degrees, PointF.Empty);
+            Matrix3x2.Invert(matrix, out matrix);
+            return matrix;
         }
 
         /// <inheritdoc/>
