+++ src/ImageSharp/Processing/Processors/Transforms/SkewProcessor.cs
@@ -50,7 +50,9 @@ namespace SixLabors.ImageSharp.Processing.Processors
         /// <inheritdoc/>
         protected override Matrix3x2 GetTransformMatrix()
         {
-            return Matrix3x2Extensions.CreateSkewDegrees(-this.DegreesX, -this.DegreesY, PointF.Empty);
+            Matrix3x2 matrix = Matrix3x2Extensions.CreateSkewDegrees(this.DegreesX, this.DegreesY, PointF.Empty);
+            Matrix3x2.Invert(matrix, out matrix);
+            return matrix;
         }
     }
 }
