+++ src/ImageSharp/Processing/Processors/Transforms/TransformProcessor.cs
@@ -30,6 +30,7 @@ namespace SixLabors.ImageSharp.Processing.Processors
         public TransformProcessor(Matrix3x2 matrix, IResampler sampler)
             : base(sampler)
         {
+            Matrix3x2.Invert(matrix, out matrix);
             this.TransformMatrix = matrix;
         }
